requirements.txt
akshare>=1.12.0
apscheduler>=3.10.0
plyer>=2.1.0
pyyaml>=6.0
requests>=2.31.0
flask>=3.0.0
